setup.py: Remove "distribute" from install_requires
Distribute was merged back into setuptools and deprecated.

But I also don't think having `install_requires` include `distribute` or
`setuptools` does anything, because it's too late. If anything, it
should go in `setup_requires` though it is very rare for folks to do
this either. The `from setuptools import setup` is usually all that is
needed.